+++ tests/testthat/test-vars-select-eval.R
@@ -46,3 +46,60 @@ test_that("data expressions support character vectors (#78)", {
   expect_identical(vars_select(letters, (identity(letters[[1]]))), vars_select(letters, a))
   expect_identical(vars_select(letters, c(identity(letters[[1]]))), vars_select(letters, a))
 })
+
+test_that("boolean operators are overloaded", {
+  expect_identical(
+    vars_select(letters, starts_with("a") & ends_with("a")),
+    vars_select(letters, intersect(starts_with("a"), ends_with("a"))),
+  )
+
+  expect_identical(
+    vars_select(letters, starts_with("a") | ends_with("c")),
+    vars_select(letters, c(starts_with("a"), ends_with("c")))
+  )
+
+  expect_identical(
+    vars_select(letters, starts_with("a") | ends_with("c") | contains("z")),
+    vars_select(letters, starts_with("a"), ends_with("c"), contains("z"))
+  )
+
+  expect_identical(
+    vars_select(letters, (starts_with("a") | ends_with("c")) & contains("a")),
+    vars_select(letters, intersect(c(starts_with("a"), ends_with("c")), contains("a")))
+  )
+
+  expect_identical(
+    vars_select(letters, !(starts_with("a") | ends_with("c"))),
+    vars_select(letters, -(starts_with("a") | ends_with("c"))),
+  )
+
+  # This pattern is not possible with `intersect()` because its
+  # arguments are evaluated in non-data context
+  expect_error(
+    vars_select(letters, intersect(c(starts_with("a"), ends_with("c")), b:d)),
+    "not found"
+  )
+  expect_identical(
+    vars_select(letters, (starts_with("a") | ends_with("c")) & b:d),
+    vars_select(letters, c)
+  )
+
+  expect_identical(
+    vars_select(letters, (starts_with("a") | ends_with("c")) | i:k),
+    vars_select(letters, c(starts_with("a"), ends_with("c")), i:k),
+  )
+})
+
+test_that("scalar boolean operators fail informatively", {
+  verify_output(test_path("outputs", "vars-select-bool-scalar-ops.txt"), {
+    vars_select(letters, starts_with("a") || ends_with("b"))
+    vars_select(letters, starts_with("a") && ends_with("b"))
+  })
+})
+
+test_that("can't use boolean operators with symbols", {
+  verify_output(test_path("outputs", "vars-select-bool-symbols.txt"), {
+    vars_select(letters, starts_with("a") & z)
+    vars_select(letters, starts_with("a") | z)
+  })
+})